data

+++ app/controllers/admin/groups_controller.rb
@@ -11,16 +11,22 @@ class Admin::GroupsController < Admin::AdminController
 
   def users
     group = Group.find(params[:group_id].to_i)
-    render_serialized(group.users.limit(100).to_a, BasicUserSerializer)
+    render_serialized(group.users.order('username_lower asc').limit(200).to_a, BasicUserSerializer)
   end
 
   def update
     group = Group.find(params[:id].to_i)
-    render_json_error if group.automatic
-    group.usernames = params[:group][:usernames]
-    group.name = params[:group][:name] if params[:name]
-    group.save!
-    render json: "ok"
+    if group.automatic
+      can_not_modify_automatic
+    else
+      group.usernames = params[:group][:usernames]
+      group.name = params[:group][:name] if params[:group][:name]
+      if group.save
+        render json: "ok"
+      else
+        render_json_error group
+      end
+    end
   end
 
   def create
@@ -33,8 +39,17 @@ class Admin::GroupsController < Admin::AdminController
 
   def destroy
     group = Group.find(params[:id].to_i)
-    render_json_error if group.automatic
-    group.destroy
-    render json: "ok"
+    if group.automatic
+      can_not_modify_automatic
+    else
+      group.destroy
+      render json: "ok"
+    end
+  end
+
+  protected
+
+  def can_not_modify_automatic
+    render json: {errors: I18n.t('groups.errors.can_not_modify_automatic')}, status: 422
   end
 end
